isolate "issues" dir from other legacy code
