+++ src/consumer/component/issues/index.ts
@@ -1,1 +1,3 @@
-export { IssuesList, IssuesClasses } from './issues-list';
+export { IssuesList, IssuesClasses, IssuesNames } from './issues-list';
+export { MISSING_DEPS_SPACE } from './component-issue';
+export { UntrackedFileDependencyEntry, MISSING_NESTED_DEPS_SPACE } from './untracked-dependencies';
